refactor: remove unecessarry
